pyproject.toml
[project]
name = "london_project"
version = "0.1.0"
description = "A template project using Python, uv, Docker and GitHub Actions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dspy>=3.1.2",
    "python-dotenv>=1",
    "redis>=7.1.0",
    "falkordb>=1.4.0",
    "docling>=1.0.0",
    "pillow>=10.0.0",
]

[project.scripts]
main = "london_project.app:main"

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"
